+++ visualization/.eslintrc.js
@@ -57,8 +57,6 @@ module.exports = {
             "off",
         "@typescript-eslint/no-explicit-any":
             "off",
-        "@typescript-eslint/no-var-requires":
-            "off",
         "@typescript-eslint/consistent-type-assertions":
             "off",
     }