+++ visualization/app/codeCharta/ui/fileChooser/fileChooser.component.ts
@@ -11,6 +11,7 @@ import { CodeChartaService } from "../../codeCharta.service"
 import { NameDataPair } from "../../codeCharta.model"
 import { StoreService } from "../../state/store.service"
 import { setIsLoadingFile } from "../../state/store/appSettings/isLoadingFile/isLoadingFile.actions"
+import zlib from "zlib"
 
 export class FileChooserController {
 	private files: NameDataPair[] = []
@@ -39,8 +40,6 @@ export class FileChooserController {
 
 				reader.onload = event => {
 					if (isCompressed) {
-						const zlib = require("zlib")
-
 						content = zlib.unzipSync(Buffer.from(event.target.result))
 					} else {
 						content = event.target.result