+++ visualization/app/codeCharta/util/treeMapHelper.ts
@@ -28,7 +28,7 @@ function buildingArrayToMap(highlighted: CodeMapBuilding[]): Map<number, CodeMap
 }
 
 function buildRootFolderForFixedFolders(map: CodeMapNode, heightScale: number, state: State, isDeltaState: boolean): Node {
-	const flattened: boolean = isNodeToBeFlat(map, state)
+	const flattened: boolean = shouldNodeBeFlat(map, state)
 	const height = FOLDER_HEIGHT
 	const width = 100
 	const length = 100
@@ -47,7 +47,7 @@ function buildRootFolderForFixedFolders(map: CodeMapNode, heightScale: number, s
 		attributes: map.attributes,
 		edgeAttributes: map.edgeAttributes,
 		deltas: map.deltas,
-		heightDelta: map.deltas?.[state.dynamicSettings.heightMetric] ? heightScale * map.deltas[state.dynamicSettings.heightMetric] : 0,
+		heightDelta: (map.deltas?.[state.dynamicSettings.heightMetric] ?? 0) * heightScale,
 		visible: isVisible(map, false, state, flattened),
 		path: map.path,
 		link: map.link,
@@ -67,7 +67,7 @@ function buildNodeFrom(
 	isDeltaState: boolean
 ): Node {
 	const isNodeLeaf = !(squaredNode.children && squaredNode.children.length > 0)
-	const flattened: boolean = isNodeToBeFlat(squaredNode.data, s)
+	const flattened: boolean = shouldNodeBeFlat(squaredNode.data, s)
 	const heightValue: number = getHeightValue(s, squaredNode, maxHeight, flattened)
 	const depth: number = squaredNode.data.path.split("/").length - 2
 	const width = squaredNode.x1 - squaredNode.x0
@@ -91,9 +91,7 @@ function buildNodeFrom(
 		attributes: squaredNode.data.attributes,
 		edgeAttributes: squaredNode.data.edgeAttributes,
 		deltas: squaredNode.data.deltas,
-		heightDelta: squaredNode.data.deltas?.[s.dynamicSettings.heightMetric]
-			? heightScale * squaredNode.data.deltas[s.dynamicSettings.heightMetric]
-			: 0,
+		heightDelta: (squaredNode.data.deltas?.[s.dynamicSettings.heightMetric] ?? 0) * heightScale,
 		visible: isVisible(squaredNode.data, isNodeLeaf, s, flattened),
 		path: squaredNode.data.path,
 		link: squaredNode.data.link,
@@ -141,18 +139,20 @@ function getOutgoingEdgePoint(width: number, height: number, length: number, vec
 	return new Vector3(vector.x - mapSize + width / 2, vector.y + height, vector.z - mapSize + 0.75 * length)
 }
 
-function isNodeToBeFlat(codeMapNode: CodeMapNode, s: State): boolean {
-	let flattened = false
-	if (s.appSettings.showOnlyBuildingsWithEdges && s.fileSettings.edges?.filter(edge => edge.visible).length > 0) {
-		flattened = nodeHasNoVisibleEdges(codeMapNode, s)
+function shouldNodeBeFlat(codeMapNode: CodeMapNode, s: State): boolean {
+	if (codeMapNode.isFlattened) {
+		return true
 	}
 
 	if (s.dynamicSettings.searchedNodePaths && s.dynamicSettings.searchPattern && s.dynamicSettings.searchPattern.length > 0) {
-		flattened = s.dynamicSettings.searchedNodePaths.size == 0 ? true : isNodeNonSearched(codeMapNode, s)
+		return s.dynamicSettings.searchedNodePaths.size === 0 || isNodeNonSearched(codeMapNode, s)
 	}
 
-	flattened = codeMapNode.isFlattened || flattened
-	return flattened
+	if (s.appSettings.showOnlyBuildingsWithEdges && s.fileSettings.edges?.filter(edge => edge.visible).length > 0) {
+		return nodeHasNoVisibleEdges(codeMapNode, s)
+	}
+
+	return false
 }
 
 function nodeHasNoVisibleEdges(codeMapNode: CodeMapNode, s: State): boolean {