+++ visualization/app/codeCharta/util/dataMocks.ts
@@ -30,6 +30,7 @@ import { ScenarioItem } from "../ui/scenarioDropDown/scenarioDropDown.component"
 import { FileSelectionState, FileState } from "../model/files/files"
 import { APIVersions, ExportCCFile } from "../codeCharta.api.model"
 import { NodeMetricDataService } from "../state/store/metricData/nodeMetricData/nodeMetricData.service"
+import packageJson from "../../../package.json"
 
 export const VALID_NODE: CodeMapNode = {
 	name: "root",
@@ -523,7 +524,7 @@ export const VALID_EDGE: Edge = {
 
 export const TEST_FILE_CONTENT: ExportCCFile = {
 	projectName: "Sample Map",
-	apiVersion: APIVersions.ONE_POINT_ONE,
+	apiVersion: APIVersions.ONE_POINT_TWO,
 	nodes: [VALID_NODE]
 }
 
@@ -537,7 +538,7 @@ export const TEST_FILE_CONTENT_INVALID_MAJOR_API = {
 export const TEST_FILE_CONTENT_INVALID_MINOR_API = {
 	fileName: "noFileName",
 	projectName: "Valid Sample Map Minor API High",
-	apiVersion: "1.2",
+	apiVersion: "1.3",
 	nodes: [VALID_NODE]
 }
 
@@ -557,7 +558,7 @@ export const TEST_FILE_CONTENT_NO_API = {
 export const FILE_META: FileMeta = {
 	fileName: "fileA",
 	projectName: "Sample Project",
-	apiVersion: "1.1"
+	apiVersion: packageJson.codecharta.apiVersion
 }
 
 export const TEST_FILE_DATA: CCFile = {
@@ -878,7 +879,7 @@ export const TEST_DELTA_MAP_A: CCFile = {
 	fileMeta: {
 		fileName: "fileA",
 		projectName: "Sample Project",
-		apiVersion: "1.1"
+		apiVersion: packageJson.codecharta.apiVersion
 	},
 	map: {
 		name: "root",
@@ -934,7 +935,7 @@ export const TEST_DELTA_MAP_B: CCFile = {
 	fileMeta: {
 		fileName: "fileB",
 		projectName: "Sample Project",
-		apiVersion: "1.1"
+		apiVersion: packageJson.codecharta.apiVersion
 	},
 	map: {
 		name: "root",
@@ -1003,7 +1004,7 @@ export const TEST_DELTA_MAP_B: CCFile = {
 }
 
 export const TEST_FILE_DATA_DOWNLOADED = {
-	apiVersion: "1.1",
+	apiVersion: packageJson.codecharta.apiVersion,
 	attributeTypes: {},
 	blacklist: [
 		{ path: "/root/bigLeaf.ts", type: "hide" },